Add XDG support
