+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+* XDG Base Directory support for application files (config, logs, data, etc.). The following directories are supported:
+  * `XDG_CONFIG_HOME` or `~/.config/zabbix-auto-config` for configuration files.
+
+### Changed
+
+* The application now looks in all valid config file locations for the config file. In order of preference:
+  * Current working directory (`./config.toml`)
+  * `XDG_CONFIG_HOME` or `~/.config/zabbix-auto-config/config.toml`
+
 ## 0.3.0
 
 ### Added
