+++ pyproject.toml
@@ -32,6 +32,7 @@ dependencies = [
     "tomli>=2.0.0",
     "packaging>=23.2",
     "typing_extensions>=4.12.0",
+    "platformdirs>=4.3.7",
 ]
